Code of FTC-Shanghai 1.18
