+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RightAuto.java
@@ -64,7 +64,7 @@ public class RightAuto extends LinearOpMode {
 //        armStretchMotor.setPower(0.5);
 //        sleep(1400);
 //        armStretchMotor.setPower(0);
-        robotTop.setTurnPosition(0.2);
+        robotTop.setTurnPosition(0.6);
         sleep(500);
 
         //lift1
@@ -100,7 +100,7 @@ public class RightAuto extends LinearOpMode {
 //        armStretchMotor.setPower(-0.5);
 //        sleep(1400);
 //        armStretchMotor.setPower(0);
-        robotTop.setTurnPosition(0);
+        robotTop.setTurnPosition(0.3);
         sleep(500);
 
         //back
