--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/PIDController.java
@@ -1,19 +1,0 @@
-package org.firstinspires.ftc.teamcode.hardware;
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.hardware.HardwareMap;
-
-import org.firstinspires.ftc.robotcore.external.Telemetry;
-
-public class PIDController {
-    OpMode opMode;
-    HardwareMap hardwareMap;
-    Telemetry telemetry;
-    public PIDController(LinearOpMode opMode) {
-        this.opMode = opMode;
-        hardwareMap = opMode.hardwareMap;
-        telemetry = opMode.telemetry;
-    }
-
-}
