+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/previousCode/Task.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.advancedManual;
+package org.firstinspires.ftc.teamcode.previousCode;
 
 public interface Task {
     int TICK_MS = 50;
