+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/previousCode/test/TurnServoTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.advancedManual.test;
+package org.firstinspires.ftc.teamcode.previousCode.test;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
