+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/AutoTest.java
@@ -1,11 +1,13 @@
 package org.firstinspires.ftc.teamcode.test;
 
 import com.qualcomm. robotcore.eventloop. opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm. robotcore. eventloop. opmode. LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm. robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.Servo;
 
+@Disabled
 @Autonomous
 public class AutoTest extends LinearOpMode {
     @Override
