+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/PIDControllerTest.java
@@ -1,11 +1,13 @@
 package org.firstinspires.ftc.teamcode.test;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.Gamepad;
 
 import org.firstinspires.ftc.teamcode.hardware.RobotTop;
 
+@Disabled
 @TeleOp
 public class PIDControllerTest extends LinearOpMode {
     @Override
