+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/ServoTest.java
@@ -6,6 +6,7 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.Servo;
 
+@Disabled
 @TeleOp(group = "Test")
 public class ServoTest extends LinearOpMode {
     @Override
