+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/WheelOffset.java
@@ -5,6 +5,7 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
+@Disabled
 @TeleOp(group = "Test")
 public class WheelOffset extends LinearOpMode {
     @Override
